Flags are not args
The following fixes a very long standing issue where flags aren't args,
but are treated as such when attempting to parse to flag set.

When the flag set encountered the non-flag it bailed causing a silent
error. This was annoying as hell and it's only now that I've realised
the error.

Instead we now correctly only tell the flagset to parse flags if it
starts with a hyphen, which in turns prevents the bailing out.